Various fixes to enable compilation on Mac OS X, and remove the glib dependency. Temporarily disables tab-completion in the drizzle client until an appropriate autoconf check can be added/enabled.

+++ include/my_global.h
@@ -634,7 +634,7 @@ typedef unsigned long uint32;
 #endif
 
 #if !defined(HAVE_ULONG) && !defined(__USE_MISC)
-typedef uint32_t	ulong;		  /* Short for unsigned long */
+typedef unsigned long ulong;		  /* Short for unsigned long */
 #endif
 #ifndef int64_t_defined
 /*
